inheritance in action, add figure class
